// src/app/api/user/helpcenter/route.ts
export type FaqItem = {
  id: string;
  category: string;
  question: string;
  answer: string;
};